Re-cut the baseline and the frozen surfaces for the wider id
`schema amend` rewrites the unshipped tip in place, so M001 keeps its number and
its name and moves only where the model did: nine column widths plus the stamp
row on PostgreSQL and SQL Server, and on SQLite the stamp row alone - its
namespaces.id was already a 64-bit AUTOINCREMENT and its namespace_id columns
already integer, so the widening asks nothing of it.

The .NET surface moves on exactly the three members that carry the id:
JobContext.NamespaceId, and NamespaceListItem's constructor and property. The
HTTP surface does not move at all, because that property is JsonIgnore - a
namespace is addressed by name over the wire and always was.

+++ anvil/Anvil.Burst/BurstDb.cs
@@ -51,7 +51,7 @@ internal sealed class BurstDb(string provider, string schema)
         var value = await command.ExecuteScalarAsync(ct);
         return value is null or DBNull
             ? throw new InvalidOperationException($"Namespace '{namespaceName}' has no row in {_prefix}namespaces yet.")
-            : Convert.ToInt16(value, CultureInfo.InvariantCulture);
+            : Convert.ToInt32(value, CultureInfo.InvariantCulture);
     }
 
     /// <summary>
